Q-learning for reinforce learning/requirements.txt
gymnasium==0.29.1
numpy>=1.21.0
matplotlib>=3.5.0
PyYAML>=6.0